tests passing for low memory mode t and f
