+++ tests/config_files/test_config_map_to_map.yaml
@@ -7,7 +7,7 @@ data:
     metadata_key: populations
     label_key: id
   ground_truth:
-    volumes: tests/data/Ground_truth/test_maps_gt_flat_10.npy
+    volumes: tests/data/Ground_truth/test_maps_gt_flat_10.pt
     metadata: tests/data/Ground_truth/test_metadata_10.csv
   mask:
     do: true
@@ -15,8 +15,8 @@ data:
 analysis:
   metrics:
     # - l2
-    # - corr
-    - corr_low_memory
+    - corr
+    # - corr_low_memory
     # - bioem
     # - fsc
     # - res
